+++ benchmark/FileIO.hs
@@ -16,8 +16,8 @@ import Data.IORef
 import Gauge
 
 import qualified Streamly.FileSystem.Handle as FH
-import qualified Streamly.Mem.Array as A
-import qualified Streamly.Mem.ArrayStream as AS
+import qualified Streamly.Memory.Array as A
+import qualified Streamly.Memory.ArrayStream as AS
 import qualified Streamly.Prelude as S
 import qualified Streamly.Fold as FL
 import qualified Streamly.String as SS