+++ src/Streamly/FileSystem/FD.hs
@@ -31,7 +31,7 @@
 --
 -- Reading and writing APIs are divided into two categories, sequential
 -- streaming APIs and random or seekable access APIs.  File IO APIs are quite
--- similar to "Streamly.Mem.Array" read write APIs. In that regard, arrays can
+-- similar to "Streamly.Memory.Array" read write APIs. In that regard, arrays can
 -- be considered as in-memory files or files can be considered as on-disk
 -- arrays.
 --
@@ -136,13 +136,13 @@ import Prelude hiding (read)
 import qualified GHC.IO.FD as FD
 import qualified GHC.IO.Device as RawIO
 
-import Streamly.Mem.Array.Types (Array(..))
+import Streamly.Memory.Array.Types (Array(..))
 import Streamly.Streams.Serial (SerialT)
 import Streamly.Streams.StreamK.Type (IsStream, mkStream)
-import Streamly.Mem.Array.Types (byteLength, defaultChunkSize)
+import Streamly.Memory.Array.Types (byteLength, defaultChunkSize)
 
 #if !defined(mingw32_HOST_OS)
-import Streamly.Mem.Array.Types (groupIOVecsOf)
+import Streamly.Memory.Array.Types (groupIOVecsOf)
 import Streamly.Streams.StreamD (toStreamD)
 import Streamly.Streams.StreamD.Type (fromStreamD)
 import qualified Streamly.FileSystem.FDIO as RawIO hiding (write)
@@ -150,8 +150,8 @@ import qualified Streamly.FileSystem.FDIO as RawIO hiding (write)
 -- import Streamly.Fold (Fold)
 -- import Streamly.String (encodeUtf8, decodeUtf8, foldLines)
 
-import qualified Streamly.Mem.Array as A
-import qualified Streamly.Mem.ArrayStream as AS
+import qualified Streamly.Memory.Array as A
+import qualified Streamly.Memory.ArrayStream as AS
 import qualified Streamly.Prelude as S
 import qualified Streamly.Streams.StreamD.Type as D